pyproject.toml
[tool.poetry]
name = "python-backend"
version = "0.1.0"
description = ""
authors = ["Gabriel Danca <gabriel.danca@endava.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
gunicorn = "^23.0.0"
alembic = "^1.18.1"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = "^2.0.45"
psycopg = "^3.3.2"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
pytest = "^9.0.2"
httpx = "^0.28.1"
pyyaml = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
